Added remove query inside the database to enable removal of task list.
Wrapped ScheduleCard with Dismissible to enable swipe action from right to left (end to start) so that it can be removed from the UI.

To remove from the database as well, onDismissed parameter was added to use another dependency injection of GetIt.I.<LocalDatabase>() and added removeSchedule() method to remove data from the database as well.

Added suffixText parameter in the customTextField to show hour text accordingly.